fix(banking): cast bigint operands in adjusted-balance INSERT — untyped $5-$6 broke reconciliation start-session with a 500
Postgres could not resolve the - operator between two unknown-typed
parameters (statement_balance_cents, beginning_balance_cents), failing
every POST /start call with 'operator is not unique... add explicit
type casts'. Both columns are bigint; casting the operands fixes the
INSERT.

+++ apps/backend/src/banking/reconciliation.routes.ts
@@ -289,7 +289,7 @@ export async function registerBankingReconciliationRoutes(app: FastifyInstance) 
               created_at,
               updated_at
             )
-            VALUES ($1,$2,$3,$4,$5,$6,$6,$5-$6,'open',now(),now())
+            VALUES ($1,$2,$3,$4,$5,$6,$6,$5::bigint-$6::bigint,'open',now(),now())
             RETURNING id
           `,
           [
